+++ java/clients/src/main/java/sleeper/clients/deploy/jar/SyncJarsRequest.java
@@ -70,6 +70,29 @@ public class SyncJarsRequest {
         return deleteOldJars;
     }
 
+    @Override
+    public int hashCode() {
+        return Objects.hash(bucketName, deploymentId, uploadFilter, Boolean.valueOf(deleteOldJars));
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (!(obj instanceof SyncJarsRequest)) {
+            return false;
+        }
+        SyncJarsRequest other = (SyncJarsRequest) obj;
+        return Objects.equals(bucketName, other.bucketName) && Objects.equals(deploymentId, other.deploymentId) && Objects.equals(uploadFilter, other.uploadFilter)
+                && deleteOldJars == other.deleteOldJars;
+    }
+
+    @Override
+    public String toString() {
+        return "SyncJarsRequest{bucketName=" + bucketName + ", deploymentId=" + deploymentId + ", uploadFilter=" + uploadFilter + ", deleteOldJars=" + deleteOldJars + "}";
+    }
+
     public static class Builder {
         private String bucketName;
         private String deploymentId;