+++ java/core/src/main/java/sleeper/core/util/cli/PositionalArgument.java
@@ -18,12 +18,11 @@ package sleeper.core.util.cli;
 /**
  * A positional argument that must be set on the command line. Used with {@link CommandArguments}.
  *
- * @param name      the name of the argument, used to retrieve it after parsing, and for use in a usage message,
- *                  e.g. if it's "arg" the message might be "Usage: command &lt;arg&gt;"
- * @param showUsage whether or not to include this argument in the usage message (usually only excluded if it is
- *                  to be passed by the system instead of the user)
+ * @param name           the name of the argument, used to retrieve it after parsing, and for use in a usage message,
+ *                       e.g. if it's "arg" the message might be "Usage: command &lt;arg&gt;"
+ * @param systemArgument whether or not this argument is passed by the system rather than the user
  */
-public record PositionalArgument(String name, boolean showUsage) {
+public record PositionalArgument(String name, boolean systemArgument) {
 
     /**
      * Creates a positional argument where its name is the same as its description in the usage message.
@@ -32,7 +31,7 @@ public record PositionalArgument(String name, boolean showUsage) {
      * @return      the argument
      */
     public static PositionalArgument create(String name) {
-        return new PositionalArgument(name, true);
+        return new PositionalArgument(name, false);
     }
 
     /**
@@ -43,7 +42,25 @@ public record PositionalArgument(String name, boolean showUsage) {
      * @return      the argument
      */
     public static PositionalArgument systemArgument(String name) {
-        return new PositionalArgument(name, false);
+        return new PositionalArgument(name, true);
+    }
+
+    /**
+     * Whether or not this argument should be included in a usage message shown to the user.
+     *
+     * @return true if it should be included
+     */
+    public boolean showUsage() {
+        return !systemArgument;
+    }
+
+    /**
+     * Whether or not this argument is set by the user.
+     *
+     * @return true if it is set by the user
+     */
+    public boolean setByUser() {
+        return !systemArgument;
     }
 
 }